PR comments

+++ onnxruntime/core/providers/cuda/cu_inc/common.cuh
@@ -129,7 +129,7 @@ template <>
 __device__ __inline__ double _Round(double a) { return rint(a); }
 
 template <>
-__device__ __inline__ half _Round(half a) { return half(rintf((float)a)); }
+__device__ __inline__ half _Round(half a) { return rint(a); }
 
 template <typename T>
 __device__ __inline__ T _Exp(T a);
